Fix a couple of warnings identified by the icc compiler

+++ programs/us_mpi_analysis/us_mpi_parse.cpp
@@ -71,7 +71,7 @@ void US_MPI_Analysis::parse_job( QXmlStreamReader& xml )
       {
          a      = xml.attributes();
          server = QHostAddress( a.value( "server" ).toString() );
-         port   = a.value( "port" ).toString().toInt();
+         port   = a.value( "port" ).toString().toUShort();
       }
 
       if ( xml.isStartElement()  &&  xml.name() == "request" )
